ChartData와 DetailData 분리
