+++ src/main/Main.java
@@ -2,6 +2,7 @@ package main;
 
 import javax.swing.*;
 import model.ChartData;
+import model.DetailData;
 public class Main {
     public static void main(String[] args){
         JFrame frame = new JFrame("The People of Music");
@@ -10,6 +11,7 @@ public class Main {
         frame.getContentPane().add(AppManager.getS_instance().getPrimaryPanel());
         AppManager.getS_instance().addToPrimaryPanel(AppManager.getS_instance().getPnlChartPrimary());
         ChartData.getS_instance();
+        DetailData.getS_instance();
 
         frame.pack();
         frame.setVisible(true);
